requirements.txt
asgiref==3.2.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.1
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-colorfield==0.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-suit==2.0a1
djangorestframework==3.11.0
future==0.18.2
idna==2.9
ndg-httpsclient==0.5.1
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
passlib==1.7.2
Pillow==7.0.0
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
serializers==0.2.4
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
